Response and request management splitup
Split up the management of the request and the endpoint with one function (others will follow). Implemented a custom exception to catch in the endpoint controller and send a custom StatusDto with more detail on errors. Currently only one endpoint uses the new functionality, which is subject to change as the new system provides more modularity. Some added functionality will also be changed as the exception thrown is not catched at a central point and may cause the programm to crash when using the newly implemented functionality to a bigger extend